솔루션 제대로 만들기
