+++ HtmlParsing/HtmlParsing.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30011.22
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HtmlParsing", "HtmlParsing\HtmlParsing.csproj", "{C6DF9FC5-53A1-4B7F-A838-B32B85888822}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C6DF9FC5-53A1-4B7F-A838-B32B85888822}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C6DF9FC5-53A1-4B7F-A838-B32B85888822}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C6DF9FC5-53A1-4B7F-A838-B32B85888822}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C6DF9FC5-53A1-4B7F-A838-B32B85888822}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {AF705C69-E9E4-44ED-9DD1-2A26B3ABEC7F}
+	EndGlobalSection
+EndGlobal
