--- html.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30011.22
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "html", "html\html.csproj", "{D0BD095D-1C79-47D8-9AF0-FC82D7CF055A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D0BD095D-1C79-47D8-9AF0-FC82D7CF055A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D0BD095D-1C79-47D8-9AF0-FC82D7CF055A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D0BD095D-1C79-47D8-9AF0-FC82D7CF055A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D0BD095D-1C79-47D8-9AF0-FC82D7CF055A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {00F0F042-16B8-4506-85EA-38FF4088EE57}
-	EndGlobalSection
-EndGlobal
